requirements.txt
mss
oauthlib
packaging
Pillow
pipenv
platformdirs
proglog
protobuf
pyasn1
pyasn1_modules
PyAutoGUI
pycodestyle
pycparser
pyflakes
PyGetWindow
PyJWT
PyMsgBox
pyperclip
PyRect
pyscreenshot
PyScreeze
python-docx
python3-openid
pytube
pytweening
pytz
PyYAML
requests
requests-oauthlib
rsa
social-auth-app-django
social-auth-core
sqlparse
stripe
tqdm
typing_extensions
tzdata
uritemplate
urllib3
virtualenv